Update 1962. Remove Stones to Minimize the Total
As mentioned in the problem statement, we have to find the minimum number of stone piles onto each other after performing piles[i]/2 operation k times. for this, we have to perform mentioned operation on the largest number that exists in the array. we can either use sorting, after doing performing a given operation we do sorting, and then again we perform the given operation to the largest element that exists in the array, which is located at the last index of the array. but doing sorting after each operation can be very time-consuming for large numbers.
instead of using this method, we can use Priority Queue, which keeps the elements in order. you can set elements by the preference of min heap or max heap. using max heap will poll the largest element every time. and you can perform a given operation.